Fix the bug of not being able to set image files with spaces
When you dropped an image file with spaces in its name, the program
was showing an error message saying that it couldn't open the file.
In the message you could see that if you dropped the imagem from
nautilus it replaced spaces in the file with %20 characters. I just
had to create a funciont in c to replace %20 charactes with spaces
again to solve the issue.